src/main/java/com/platform/parent/config/NewsConfig.java
package com.platform.parent.config;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.context.annotation.PropertySource;
import org.springframework.stereotype.Component;

import java.util.List;

/**
 * Created by tqyao.
 */
@Component
@ConfigurationProperties( prefix = "user")
@PropertySource("classpath:config/news.properties")
public class NewsConfig {
    private List<String> name;

    public List<String> getName() {
        return name;
    }

    public void setName(List<String> name) {
        this.name = name;
    }
}
